refactor(admin): split all creators table state handling
The all-creators table had accumulated persistence, merge,
and action orchestration logic in one client component.
This refactor moves override storage/merge behavior and
approve-reject mutation handling into focused hooks so the
table stays readable and easier to extend without changing
behavior.

Constraint: Admin still compensates for server responses
that only return pending requests
Rejected: Leave logic inline in page component |
would keep feature flow hard to reason about
Confidence: high
Scope-risk: narrow
Reversibility: clean
Directive: Keep table component focused on composition
and rendering; move future workflow state into hooks
Tested: pnpm --filter admin exec tsc --noEmit;
pnpm --filter admin lint
Not-tested: Manual browser validation after refactor
