+++ apps/admin/components/features/all-creators/AllCreatorsTable.tsx
@@ -1,82 +1,32 @@
 "use client";
 
-import { useEffect, useMemo, useState } from "react";
-import toast from "react-hot-toast";
+import { useState } from "react";
 import type { CreatorRequest } from "../../../types/creator-request";
-import {
-  useApproveCreator,
-  useCreatorRequests,
-  useRejectCreator,
-} from "../../../hooks/api";
+import { useCreatorRequests } from "../../../hooks/api";
 import { usePagination } from "../../../hooks/ui/use-pagination";
 import { AllCreatorsPanel, AllCreatorsState } from "./AllCreators.styles";
 import { CreatorRequestsTable } from "./CreatorRequestsTable";
 import { CreatorPagination } from "./CreatorPagination";
 import { CreatorDetailsModal } from "./CreatorDetailsModal";
-
-const CREATOR_REQUEST_OVERRIDES_STORAGE_KEY = "admin-creator-request-overrides";
-
-function loadCreatorOverrides(): CreatorRequest[] {
-  if (typeof window === "undefined") {
-    return [];
-  }
-
-  try {
-    const rawValue = window.localStorage.getItem(
-      CREATOR_REQUEST_OVERRIDES_STORAGE_KEY,
-    );
-
-    if (!rawValue) {
-      return [];
-    }
-
-    const parsedValue = JSON.parse(rawValue) as CreatorRequest[];
-    return Array.isArray(parsedValue) ? parsedValue : [];
-  } catch {
-    return [];
-  }
-}
+import { useCreatorRequestActions } from "./useCreatorRequestActions";
+import { useCreatorRequestOverrides } from "./useCreatorRequestOverrides";
 
 export function AllCreatorsTable() {
   const [selectedCreator, setSelectedCreator] = useState<CreatorRequest | null>(
     null,
   );
-  const [creatorOverrides, setCreatorOverrides] = useState<CreatorRequest[]>(
-    () => loadCreatorOverrides(),
-  );
   const creatorRequestsQuery = useCreatorRequests();
-  const approveCreatorMutation = useApproveCreator();
-  const rejectCreatorMutation = useRejectCreator();
-
-  useEffect(() => {
-    if (typeof window === "undefined") {
-      return;
-    }
-
-    window.localStorage.setItem(
-      CREATOR_REQUEST_OVERRIDES_STORAGE_KEY,
-      JSON.stringify(creatorOverrides),
-    );
-  }, [creatorOverrides]);
-
-  const creators = useMemo(() => {
-    const serverCreators = creatorRequestsQuery.data ?? [];
-    const creatorMap = new Map<string, CreatorRequest>();
-
-    for (const creator of serverCreators) {
-      creatorMap.set(creator.id, creator);
-    }
-
-    for (const creator of creatorOverrides) {
-      creatorMap.set(creator.id, creator);
-    }
-
-    return Array.from(creatorMap.values()).sort(
-      (firstCreator, secondCreator) =>
-        new Date(secondCreator.createdAt).getTime() -
-        new Date(firstCreator.createdAt).getTime(),
-    );
-  }, [creatorOverrides, creatorRequestsQuery.data]);
+  const { creators, updateCreatorStatus } = useCreatorRequestOverrides(
+    creatorRequestsQuery.data ?? [],
+  );
+  const {
+    activeAction,
+    activeRequestId,
+    handleApproveCreator,
+    handleRejectCreator,
+  } = useCreatorRequestActions({
+    onCreatorUpdated: (creator) => updateCreatorStatus(creator, creator.status),
+  });
 
   const totalItems = creators.length;
   const pagination = usePagination({
@@ -84,81 +34,6 @@ export function AllCreatorsTable() {
     totalItems,
     initialPageSize: 10,
   });
-
-  const activeAction = approveCreatorMutation.isPending
-    ? "approve"
-    : rejectCreatorMutation.isPending
-      ? "reject"
-      : null;
-
-  const activeRequestId =
-    approveCreatorMutation.variables?.requestId ??
-    rejectCreatorMutation.variables?.requestId ??
-    null;
-
-  const handleApproveCreator = (creator: CreatorRequest) => {
-    approveCreatorMutation.mutate(
-      { requestId: creator.id },
-      {
-        onSuccess: () => {
-          const updatedCreator: CreatorRequest = {
-            ...creator,
-            status: "approved",
-            updatedAt: new Date().toISOString(),
-          };
-
-          setCreatorOverrides((currentOverrides) => {
-            const nextOverrides = currentOverrides.filter(
-              (currentCreator) => currentCreator.id !== creator.id,
-            );
-
-            return [...nextOverrides, updatedCreator];
-          });
-
-          if (selectedCreator?.id === creator.id) {
-            setSelectedCreator(updatedCreator);
-          }
-
-          toast.success("Creator approved successfully");
-        },
-        onError: (error) => {
-          toast.error(error.message || "Failed to approve creator");
-        },
-      },
-    );
-  };
-
-  const handleRejectCreator = (creator: CreatorRequest) => {
-    rejectCreatorMutation.mutate(
-      { requestId: creator.id },
-      {
-        onSuccess: () => {
-          const updatedCreator: CreatorRequest = {
-            ...creator,
-            status: "rejected",
-            updatedAt: new Date().toISOString(),
-          };
-
-          setCreatorOverrides((currentOverrides) => {
-            const nextOverrides = currentOverrides.filter(
-              (currentCreator) => currentCreator.id !== creator.id,
-            );
-
-            return [...nextOverrides, updatedCreator];
-          });
-
-          if (selectedCreator?.id === creator.id) {
-            setSelectedCreator(updatedCreator);
-          }
-
-          toast.success("Creator rejected successfully");
-        },
-        onError: (error) => {
-          toast.error(error.message || "Failed to reject creator");
-        },
-      },
-    );
-  };
 
   if (creatorRequestsQuery.isLoading) {
     return (
@@ -192,8 +67,20 @@ export function AllCreatorsTable() {
       <CreatorRequestsTable
         creators={pagination.paginatedData}
         onSelectCreator={(creator) => setSelectedCreator(creator)}
-        onApproveCreator={handleApproveCreator}
-        onRejectCreator={handleRejectCreator}
+        onApproveCreator={(creator) =>
+          handleApproveCreator(creator, (updatedCreator) => {
+            if (selectedCreator?.id === creator.id) {
+              setSelectedCreator(updatedCreator);
+            }
+          })
+        }
+        onRejectCreator={(creator) =>
+          handleRejectCreator(creator, (updatedCreator) => {
+            if (selectedCreator?.id === creator.id) {
+              setSelectedCreator(updatedCreator);
+            }
+          })
+        }
         activeAction={activeAction}
         activeRequestId={activeRequestId}
       />
